refactor(ui): rename ui component names to kebab case
Shadcn recommendations
