--- components/index.ts
@@ -1,11 +1,0 @@
-
-// UI
-export { Button } from "./shared/ui/Button/Button";
-export * from "./shared/ui/DropdownMenu/DropdownMenu";
-export { default as Footer } from "./shared/ui/Footer/Footer";
-export { default as Header } from "./shared/ui/Header/Header";
-export { Input } from "./shared/ui/Input/Input";
-export { default as Loading } from "./shared/ui/Loading/Loading";
-export { ModeToggle } from "./shared/ui/ModeToggle/ModeToggle";
-export { Textarea } from "./shared/ui/Textarea/Textarea";
-
