+++ components/ui/index.ts
@@ -1,0 +1,9 @@
+export { Button } from "./button/button";
+export * from "./dropdown-menu/dropdown-menu";
+export { default as Footer } from "./footer/footer";
+export { default as Header } from "./header/header";
+export { Input } from "./input/input";
+export { default as Loading } from "./loading/loading";
+export { ModeToggle } from "./mode-toggle/mode-toggle";
+export { Textarea } from "./textarea/textarea";
+
